pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
only-packages = true

[project]
name = "lmxy"
version = "0.1.10"
description = "Language models extensions"
readme = "README.md"
requires-python = ">=3.12"
license = {file = "LICENSE"}
keywords = []
authors = [
    {name = "Paul Maevskikh", email = "arquolo@gmail.com"},
]
maintainers = [
    {name = "Paul Maevskikh", email = "arquolo@gmail.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "httpx",
    "pydantic~=2.10",
    "pydantic-settings",
    "tenacity",
]

[project.urls]
homepage = "https://github.com/arquolo/lmx"

[project.optional-dependencies]
base = [  # responses, rerank, embed
    "llama-index-core~=0.13.0",
    "llama-index-utils-huggingface",
]
tokenizer = [
    "glow~=0.15.0",
    "tiktoken",
    "transformers",
]
fastembed = [
    "fastembed~=0.7.0",
    "glow~=0.15.0",
]
openailike = [
    "lmxy[base,tokenizer]",
    "openai~=1.102",
]
qdrant = [
    "lmxy[base,fastembed]",
    "grpcio",  # transitive
    "orjson",
    "qdrant-client~=1.15.1",
]
all = [
    "lmxy[openailike,qdrant]",
]
dev-core = [
    "black~=25.1",
    "flake8~=7.0",
    "flake8-pie",
    "flake8-pyi",
    "flake8-pyproject",
    "flake8-simplify",
    "isort",
    "mypy~=1.17",
    "pytest~=8.0",
    "pytest-asyncio",
    "ruff~=0.12.2",
    "types-requests",
]
dev = [
    "lmxy[dev-core]",
    "flake8-alphabetize",
    # "flake8-class-attributes-order",
    # "flake8-newspaper-style",
]
dev-wemake = [
    "lmxy[dev-core]",
    "wemake-python-styleguide~=1.3.0",
]

# Lint

[tool.ruff]
line-length = 79
target-version = "py312"

[tool.ruff.lint]
select = [
    "A", "B", "C4", "C9", "D", "E", "F", "N", "PT", "Q0", "RET", "S", "SIM", "UP", "W",
    "BLE", "COM", "ICN", "INP", "PL", "RSE", "RUF",
    # "FBT", "SLF", "TCH",
    "TRY",
    "NPY201",
]
ignore = [
    "A003", "B008", "B905", "D1", "D2", "D4", "N812", "S1", "S3",
    "COM812", "COM818", "PLR0913", "PLR2004", "PT011", "RET504", "TRY003", "SIM105",
]
exclude = [".eggs", ".git", "__pycache__", "build", "dist"]
unfixable = ["I001"]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.format]
quote-style = "single"

[tool.flake8]
select = [
    # pyflakes,pycodestyle: n/a in ruff
    "E", "F703", "F721", "F723", "F831", "W",
    # flake8-alphabetize
    "AZ400",
    # flake8-class-attributes-order: CCE{001-002}
    # "CCE",
    # flake8-newspaper-style: NEW100
    # "NEW",
    # flake8-pie
    "PIE787", "PIE789", "PIE793", "PIE795", "PIE798", "PIE799", "PIE805", "PIE806",
    # flake8-pyi: Y{002-006,009-026,028-052}
    "Y0",
    # flake8-simplify
    "SIM113", "SIM116", "SIM9",
    # wemake-python-styleguide
    "WPS",
]
ignore = [
    "E203", "E302", "E704",  # black fmt
    "W503", "W504",
    "WPS1",
    "WPS2",
    "WPS300", "WPS301", "WPS303", "WPS304", "WPS305", "WPS306", "WPS309", "WPS314", "WPS316", "WPS317", "WPS318", "WPS319", "WPS326", "WPS330", "WPS331", "WPS332", "WPS335", "WPS336", "WPS337", "WPS338", "WPS339", "WPS347", "WPS349", "WPS352", "WPS356", "WPS358", "WPS359", "WPS361", "WPS362",
    "WPS407", "WPS410", "WPS412", "WPS414", "WPS420", "WPS421", "WPS424", "WPS425", "WPS428", "WPS429", "WPS430", "WPS431", "WPS432", "WPS433", "WPS434", "WPS435", "WPS436", "WPS437", "WPS440", "WPS441", "WPS442", "WPS444", "WPS450", "WPS451", "WPS457", "WPS458", "WPS459", "WPS460", "WPS461", "WPS465",
    "WPS501", "WPS503", "WPS509", "WPS526", "WPS527", "WPS529",
    "WPS600", "WPS601", "WPS602", "WPS603", "WPS604", "WPS609",
]
exclude = [".eggs", ".git", "__pycache__", "build", "dist"]
max-complexity = 10

# Type check

[tool.mypy]
python_version = "3.12"
sqlite_cache = true
check_untyped_defs = true
disallow_subclassing_any = true
extra_checks = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
strict_equality = true
warn_redundant_casts = true
warn_unused_ignores = true
exclude = ["dist"]

# Format

[tool.black]
line-length = 79
target-version = ['py312']
skip-string-normalization = true
